Module google_coordinate1::api

Structs

Coordinate
Central instance to access all Coordinate related resource activities

CustomField
Custom field.

CustomFieldDef
Custom field definition.

CustomFieldDefListCall
Retrieves a list of custom field definitions for a team.

CustomFieldDefListResponse
Collection of custom field definitions for a team.

CustomFieldDefMethods
A builder providing access to all methods supported on customFieldDef resources. It is not used directly, but through the Coordinate hub.

CustomFields
Collection of custom fields.

EnumItemDef
Enum Item definition.

Job
A job.

JobChange
Change to a job. For example assigning the job to a different worker.

JobGetCall
Retrieves a job, including all the changes made to the job.

JobInsertCall
Inserts a new job. Only the state field of the job should be set.

JobListCall
Retrieves jobs created or modified since the given timestamp.

JobListResponse
Response from a List Jobs request.

JobMethods
A builder providing access to all methods supported on job resources. It is not used directly, but through the Coordinate hub.

JobPatchCall
Updates a job. Fields that are set in the job state will be updated. This method supports patch semantics.

JobState
Current state of a job.

JobUpdateCall
Updates a job. Fields that are set in the job state will be updated.

Location
Location of a job.

LocationListCall
Retrieves a list of locations for a worker.

LocationListResponse
Response from a List Locations request.

LocationMethods
A builder providing access to all methods supported on location resources. It is not used directly, but through the Coordinate hub.

LocationRecord
Recorded location of a worker.

Schedule
Job schedule.

ScheduleGetCall
Retrieves the schedule for a job.

ScheduleMethods
A builder providing access to all methods supported on schedule resources. It is not used directly, but through the Coordinate hub.

SchedulePatchCall
Replaces the schedule of a job with the provided schedule. This method supports patch semantics.

ScheduleUpdateCall
Replaces the schedule of a job with the provided schedule.

Team
A Coordinate team.

TeamListCall
Retrieves a list of teams for a user.

TeamListResponse
Response from a List Teams request.

TeamMethods
A builder providing access to all methods supported on team resources. It is not used directly, but through the Coordinate hub.

TokenPagination
Pagination information.

Worker
A worker in a Coordinate team.

WorkerListCall
Retrieves a list of workers in a team.

WorkerListResponse
Response from a List Workers request.

WorkerMethods
A builder providing access to all methods supported on worker resources. It is not used directly, but through the Coordinate hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.