Struct google_coordinate1::CustomFieldDefMethods

pub struct CustomFieldDefMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on customFieldDef resources. It is not used directly, but through the Coordinate hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_coordinate1 as coordinate1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use coordinate1::Coordinate;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Coordinate::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `list(...)`
// to build up your call.
let rb = hub.custom_field_def();

Methods

impl<'a, C, A> CustomFieldDefMethods<'a, C, A>

fn list(&self, team_id: &str) -> CustomFieldDefListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of custom field definitions for a team.

Arguments

teamId - Team ID