Struct google_coordinate1::JobMethods

pub struct JobMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on job resources. It is not used directly, but through the Coordinate hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_coordinate1 as coordinate1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use coordinate1::Coordinate;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Coordinate::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.jobs();

Methods

impl<'a, C, A> JobMethods<'a, C, A>

fn get(&self, team_id: &str, job_id: &str) -> JobGetCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a job, including all the changes made to the job.

Arguments

teamId - Team ID

jobId - Job number

fn update(&self, request: Job, team_id: &str, job_id: &str) -> JobUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a job. Fields that are set in the job state will be updated.

Arguments

request - No description provided.

teamId - Team ID

jobId - Job number

fn patch(&self, request: Job, team_id: &str, job_id: &str) -> JobPatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates a job. Fields that are set in the job state will be updated. This method supports patch semantics.

Arguments

request - No description provided.

teamId - Team ID

jobId - Job number

fn list(&self, team_id: &str) -> JobListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves jobs created or modified since the given timestamp.

Arguments

teamId - Team ID

fn insert(&self, request: Job, team_id: &str, address: &str, lat: f64, lng: f64, title: &str) -> JobInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new job. Only the state field of the job should be set.

Arguments

request - No description provided.

teamId - Team ID

address - Job address as newline (Unix) separated string

lat - The latitude coordinate of this job's location.

lng - The longitude coordinate of this job's location.

title - Job title