Struct google_coordinate1::JobUpdateCall

pub struct JobUpdateCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Updates a job. Fields that are set in the job state will be updated.

A builder for the update method supported by a job resource. It is not used directly, but through a JobMethods instance.

Example

Instantiate a resource method builder

use coordinate1::Job;
 
// As the method needs a request, you would usually fill it with the desired information
// into the respective structure. Some of the parts shown here might not be applicable !
// Values shown here are possibly random and not representative !
let mut req = Job::default();
 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.jobs().update(req, "teamId", "jobId")
             .title("et")
             .progress("duo")
             .note("aliquyam")
             .lng(0.915099997853)
             .lat(0.45715758654)
             .customer_phone_number("eos")
             .customer_name("erat")
             .add_custom_field("sadipscing")
             .assignee("dolor")
             .address("eirmod")
             .doit();

Methods

impl<'a, C, A> JobUpdateCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, Job)>

Perform the operation you have build so far.

fn request(self, new_value: Job) -> JobUpdateCall<'a, C, A>

Sets the request property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn team_id(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Team ID

Sets the team id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn job_id(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Job number

Sets the job id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn title(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Job title

Sets the title query property to the given value.

fn progress(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Job progress

Sets the progress query property to the given value.

fn note(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Job note as newline (Unix) separated string

Sets the note query property to the given value.

fn lng(self, new_value: f64) -> JobUpdateCall<'a, C, A>

The longitude coordinate of this job's location.

Sets the lng query property to the given value.

fn lat(self, new_value: f64) -> JobUpdateCall<'a, C, A>

The latitude coordinate of this job's location.

Sets the lat query property to the given value.

fn customer_phone_number(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Customer phone number

Sets the customer phone number query property to the given value.

fn customer_name(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Customer name

Sets the customer name query property to the given value.

fn add_custom_field(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Sets the value of custom fields. To set a custom field, pass the field id (from /team/teamId/custom_fields), a URL escaped '=' character, and the desired value as a parameter. For example, customField=12%3DAlice. Repeat the parameter for each custom field. Note that '=' cannot appear in the parameter value. Specifying an invalid, or inactive enum field will result in an error 500.

Append the given value to the custom field query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn assignee(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Assignee email address, or empty string to unassign.

Sets the assignee query property to the given value.

fn address(self, new_value: &str) -> JobUpdateCall<'a, C, A>

Job address as newline (Unix) separated string

Sets the address query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> JobUpdateCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> JobUpdateCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> JobUpdateCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Full.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.